1.optimal_filter.py path refactoring changes, 2.changes required for VM telemetry API end point
